810: Reihenfolge der Artikelpositionen nicht konstant
Task-Url: https://bugs.fakturama.info/view.php?id=810
- fixed
- added some null checks

+++ Fakturama-Parent/com.sebulli.fakturama.rcp/src/main/java/com/sebulli/fakturama/parts/ProductEditor.java
@@ -366,7 +366,7 @@ public class ProductEditor extends Editor<Product> {
             part.setLabel(msg.commandNewProductName);
 
 			// Set the vat to the standard value
-            int vatId = defaultValuePrefs.getInt(Constants.DEFAULT_VAT);
+            long vatId = defaultValuePrefs.getLong(Constants.DEFAULT_VAT);
             VAT vat = vatDao.findById(vatId);  // initially set default VAT
 			editorProduct.setVat(vat);
